Support servers that URL decode requests and those that don't
Previously we assumed that if Canopy made a request for a static file
with URL encoding eg `100%25.json`, the server would check for a file
on disk called "100%.json", and so, the `canopy build` command would
URL encode topics once to produce filenames(eg, `100%` ->
`100%25.json`), but the front-end code would encode the original topic
name twice in the request (eg `100%2525.json`) so that once the server
decoded once, it still resulted in the once-encoded filename.

However, some hosting providers like Netlify do not decode static asset
requests, and so for them, we should be requesting the once-encoded
asset name (ie `100%25.json`.) So, in order to support both types of
providers, we first make a request for the singly encoded version, eg
`100%25.json`, and if that fails, we try the doubly-encoded version ie
`100%2525.json` before erroring.

This order of trials is safe because there shouldn't be an issue of
collisions in this direction, ie, if we try `100%2525.json` first, we
are either hitting the file `100%25.json` (for a server that decodes, )
corresponding to an original topic name of "100%", or, we are hitting
the file `100%2525.json` (for a server that doesn't encode),
corresponding to the original topic name "100%25", which is an unlikely
but possible value, therefore if both topics exist in a project,
requesting `100%2525.json` would return different files depending on
whether the hosting provider decodes requests, and we might
accidentally get the wrong asset.

However, if we try `100%25.json` first, we are either hitting the file
`100%.json` (for a server that decodes) or `100%25.json` (for a server
that doesn't,) and, we know that there is no topic name that would
produce the filename `100%.json` because all percent signs are at least
once-encoded to produce the filename string. Therefore, it is safe to
try `100%25.json` without fear of collision, and if we don't find the
file, we can try the double-encoded version, knowing it is possible
that we might be accidentally getting the JSON corresponding to an
original topic name of "100%25" that has been once-encoded to produce
the filename `100%2525.json`, as opposed to topic "100%" that has been
twice-encoded to produce the request filename `100%2525.json`, etc.