Fix page jump when refreshing on scrolled page
Previously, if the user loaded a page at an existing path, the page
would render without a scroll, and then there would be a visible jump
as the page scrolled to the correct position.

Now, for initial page loads (ie, where the display option scrollStyle is
set to 'auto'), instead of displaying the paragraphs of the path,
we set the opacity to 0 so that they are not visible, but we set display
to 'block' so that they start taking up space, then, we scroll to the
correct spot on the blank page, and only then do we set opacity to 1,
causing the content to become visible at the correct scroll position.