Hoist category topic in bulk file
Often, one creates a category like Todos that contains a topic Todos
which then links to several other topics. Previously, the bulk file
would render the topics in alphabetical order even if one had the
same name as the category and so should obviously be first, and so one
would have to create a container category like [Todos/Subcategories]
and put only the topic "Todos" in the "Todos" category.

For purposes of this comparison we ignore parenthetical text so that the
category can be eg [Restaurants/New York] and the topic can be
"New York (Restaurants)" and it will get hoisted.