Arrow key refactor
Previously, the left and right keys cycled through the links of the
current paragraph while down selected a link and up returned to parent.

This was a simple system, but created the confusing effect that up
and down within a paragraph that contained vertically stacked links
did not operate in the way most users expected it to, ie changing
paragraphs as opposed to navigating among the links of the given
paragraph.

This commit introduces a new system for arrow key handling wherein
changes in selection are calculated based on the position of the old
and new link in 2D space. This produces a large increase in code
complexity, but in most cases better approximates the user's intuitions
about arrow key behavior. One exception remains that when the user
presses down, it is never clear whether they mean to go to a lower link
in the paragraph, or to select the current link and go to its child
paragraph. For the time being, "down" will do the former, navigating
links of the given paragraph, whereas return will go to the child
paragraph, however, this may still cause confusion among some users.

The feature includes support for wrapping, from the end of a line
to the start of the next, and from the end of a paragraph to the
start of that paragraph.

A large part of the complexity of the feature is due to support for
right-to-left text, ie, when pressing left or right at the end of a
line or paragraph, which direction to wrap (horizontally and
vertically) depends on the orientation of the text (both that of the
current link, and of the next link.)

This commit also changes the behavior of command click, return, and
related link selection commands. Previously, it was assumed the user
wanted to reset the current path to the new topic unless stated
otherwise. Now, we make the opposite presumption, that return/new tab
mean to inline the new topic at the current path unless otherwise
specified using the alt key.