Arrow key refactor
Previously, the left and right keys cycled through the links of the
current paragraph while "down" selected a link and "up" returned to the
parent paragraph.

This was a simple system, but created the confusing effect that "up"
and "down" within a paragraph that contained vertically-stacked links
did not behave as most users expected, burrowing into a child paragraph
as opposed to iterating through the links of the current paragraphs.

This commit introduces a new system for arrow key handling wherein
changes in selection are calculated based on the position of the old
and new link in 2D space. This produces a large increase in code
complexity, but in most cases better approximates the user's intuitions
about arrow key behavior. One exception remains that when the user
presses down, it is never clear whether they intend to go to a lower
link in the same paragraph, or to select the current link and go to its
child paragraph. For the time being, "down" will do the former,
navigating links of the given paragraph, whereas return will go to the
child paragraph. However, this may still cause confusion among some
users.

This feature includes support for wrapping, from the end of a line
to the start of the next, and from the end of a paragraph to the
start of that paragraph.

A large part of the complexity of the feature is due to support for
right-to-left text, ie, when pressing left or right at the end of a
line or paragraph, which direction to wrap (horizontally and
vertically) depends on the orientation of the text (both that of the
current link, and that of the subsequent link.)

Additionally, "up", "down", and "enter" when the given link is off
screen will no longer scroll directly to the new position, but will
rather scroll up to a fixed amount, requiring the user to press
iteratively in order to reach the desired location. This has the benefit
of allowing long paragraphs to be viewed with the arrow keys alone,
which wasn't previously possible, and has the effect of keeping the user
spatially oriented.

The automatic scrolling behavior when link selection changes is now
sensitive to the manner in which the link was selected, so, if a link
is clicked, the child paragraph will be more aggressively shown,
because the user's desire to select that link is clearer, as opposed to
key-stroke-initiated link selections, which do not so clearly indicate
a desire to scroll to the child paragraph, because the user may just be
iterating through that link to get to a different one, and so we want
to retain focus on the link's enclosing paragraph until the user
indicates otherwise by preforming a repetitive selective action.

This commit also changes the behavior of command-click, return, and
related link selection commands. Previously, it was assumed the user
wanted to reset the current path to the new topic unless they stated
otherwise. Now, we make the opposite presumption, that return/command
means to inline the new topic at the current path unless the user
overrides this behavior using the alt key.