Variable size table lists
Previously, all table lists had cells of fixed size, and if there was
too much text to fit, the font-size was decreased. Now, we look at the
total number of characters in the cell and the length of the longest
word which constrains width and use that to choose the proper
size for the table list.