Bug: Table lists with invalid ends cause catastrophic backtracking
Previously we matched table lists by looking for a delimiter, bullet
lines, ending with a delimiter, but in the case where there is an
opening delimiter (or empty bullet that counts as such), the regex
will try every substring of the table list before giving up. Now,
we check that there is a valid end delimiter before validating the
content.