Back button functionality
Users are encouraged in some circumstances to make deep linear chains of
subtopics in order to communicate a point in steps to the reader.
However, this and other similar situations results in the user being
very far down the page and desiring to return to the paragraph that
linked to the topic of the given subtopic chain, and so this commit
introduces a "back button" which is an absolutely positioned element
at the bottom of the page that appears when the user has scrolled a
certain distance beyond the current bottom of the page.

This commit also disables a behavior where image tokens or images in
html tokens would rescroll the page on-load to ensure that the loading
didn't disrupt the viewport, this behavior may be useful in future but
for the time-being it is causing problems for the playwright browser
tests because they trigger a scroll much later after the test runner
has already moved on in the page interactions. Ideally we might check
if the user has interacted with the page since the call to displayPath,
and if yes, cancel these on-load scrolls.