Offer --use-existing bulk file option for bulk --sync mode
Previously canopy bulk sync mode would create a new bulk file even if
there was one already present in the project directory. However, if the
user's sync process was killed and they continued writing to the bulk
file unwittingly, this could cause data loss. Additionally, if the user
is using the canopyd development binary, every code change will trigger
a restart and regeneration of the bulk file, and it is theoretically
possible for recent changes to the bulk file not to have made it to the
topic files for some reason, and thus those changes would be lost.

Perhaps this change should be made to the --start option also, however,
in that case the user seems more likely to be aware of the state of
their bulk file and has the option to run --finish if they want to adopt
the extant bulk file.