Path references
This commit introduces a series of wide-ranging changes that remove
the concept of import references and introduce the concept of path
references, merging them into the concept of global references. Now,
all global references reference a path of one or more paragraphs.
Path references can introduce cycles and trigger the recently introduced
cycle reduction behavior.

- Back button is removed because now the author can create customized
back-buttons at the data level, ie path references that refer to the
topic subtopic are effectively back-references.

- UpdateView now handles slow network connections, updating the UI while
waiting for data from the network.

- The decision to animate is now merely a function of the distance
up we are going, and whether there is a second leg to the transition.

- Image load handling is improved. Now, images have a fixed height,
and when they load, we pick a focal element and maintain it's position
pre- and post- load.

- Previously there was a concept of selectALink wherein a caller to
updateView could specify the selection of an as-yet-non-extant link
upon render, but due to the introduction of promise handles on
updateView and its various wrappers, plus the new renderOnly option,
such link selection is now falling on the caller to specify manually
as in click handlers, rather than on the library to support as an option.

- Many tests are updated or added to handle new behaviors