Add --cache build option
When projects reach the 50K word mark, build times are getting to be
1-2 seconds, which is noticeable if someone is using the bulk --sync
option.

Now, users can invoke `build`, `watch`, and `bulk --sync` with the
`--cache` option. This causes the build to be done in two steps,
first, building only those expl files that have been changed, then,
building all expl files. The latter pass is necessary as a safe-guard
because we can't 100% rely on the fact that a expl file is old in
deciding to ignore it, because for example an old expl file might
contain a path reference that is predicated on adjacencies in other
topic files that may have changed, and so it is dangerous not to parse
all expl files on a build because then such errors will not be detected,
but for sync users it will make a very noticeable difference to build
the new JSON files on presumption of a good build, and to only error
afterwords if it was found not to be.