Add client performance features
In conjunction with the previously committed `--cache` option that
improves support for large data sets, this commit introduces several
performance-related features on the client side.

First, a bug is fixed which caused eager rendering to target links
that were not visible. Now, links will have a `renderOnly`
request initiated when they become visible.

Secondly, on each path change, we now "prune" the DOM, removing
any paragraph that is not visible or targeted by a visible link.

Lastly, pure functions like removeStyleTokens or Topic#constructor
that are heavily used are now cached to save time.