Cache fetch requests
Eager render causes us to make several overlapping requests around the
same time, which means that the same paragraph might get constructed
multiple times to accommodate different requested paths.

Already we had been caching JSON requests, so that if a second request
for the same file was made, it received the same promise as the first
one.

However, it was possible for two requests to get copies of that promise
and both plan to construct section elements from it before either
completed, causing duplicate paragraphs to be created. Now, we cache
open requests for paragraphs and reuse them for different path requests.