Shadow DOM implementation
Previously all content including eagerly-rendered content was stored in
the DOM and shown or not shown depending on the current selection.
However, for large topics or large paths or heavy eager-rendering, the
size of the DOM could produce a noticeable UI delay. So, now the
Paragraph class is expanded to store information about the hierarchy of
paragraphs, and to reflect that hierarchy or not reflect it in the
actual DOM as needed per-request.

As part of this feature, an effort is made when initializing a paragraph
class object for a section element to preserve and recycle the same
instance per section element throughout the app life-cycle.