Decode encoded pound signs & revamp encoding scheme
When doing path parsing in Path#stringToArray or in
Topic#fromEncodedSlug (now Topic#fromUrl), we have to decide which
characters that seem encoded are really encoded characters, and which
are to be interpreted as literal strings eg "%23". In general, the
approach has been to presume strings that look encoded in a URL are,
and to decode them, unless they are preceded by an encoded backslash
(%5C), and conversely, when decoding, to remove encoded backslashes
once the character has been identified as literal.

For example, when we see an underscore in a URL, we may not know whether
it was originally a literal underscore, or whether it was a space that
became an underscore during URL encoding. When we see %23, we don't
know whether it was a # in a topic name like "#1" that became a %23 in
URL encoding, whether it is a # from the path structure that got
accidentally encoded, or whether it is the literal letters %23 in a topic
name.

Things might be simpler if we just URL encoded all characters, but in
order to have human-readable URLs, we want to encode the minimum number
of characters to get valid, reversible URLs, while leaving any
unproblematic characters as they are.

When designing a minimal URL encoding scheme, there are a few types of
cases to consider. For example, looking at the "&" character:

`&`     - this might not be a valid URL character
`%26`   - this is the character "&"
`%2526` - these are the literal characters "%26" because escaped

Greater difficulty exists when there is a character whose the encoded
version gets converted to the non-encoded version by the browser before
JS execution, preventing us from using the encoded form to represent
the literal character.

` `    - this would never show up in a URL
`_`    - if this is in a URL, it was previously a space
`%5F`  - this was a literal _, but it gets swapped to `_` before JS, so
         will be indistinguishable to JS from the previous `_` and thus
         cannot be our encoded representation of literal _.
`%5C_` - We are forced
         to use this to represent a literal underscore.

Motivating this commit, pound signs ie # present a similar problem.
First, pound signs present the same problem as all URL-invalid
characters, that they need to be encoded to a unique encoding that is
distinguishable from all other string sequences. For this concern,
theoretically regular URL encoding should suffice, ie using %23 for a
pound in a topic name, and using %2523 for a literal `%23` in a topic
name.

Now, how can we tell which pound signs are part of topic names and
which ones are part of path structure eg /A#B/C#D ? Theoretically, we
should have been able to just say that a real # in a URL is part of the
path structure, and that any %23 was a pound in a topic name.

So, we might have thought that this encoding scheme should suffice:

`#`     - this in a URL is a fragment id ie subtopic marker eg A#B/C#D
`%23`   - this is a # that was part of a topic name
`%2523` - this is a literal `%23` that was part of a topic name

Writing "\#" was necessary within expl file references when using a
pound sign literal in a topic name, however, once the parser has
distinguished between a topic-name-# and a subtopic-marker-#, when
crafting the URL, regular URL encoding should suffice to distinguish
between the two, ie, the # that was escaped in the `.expl` link text
becomes %23 while the path-subtopic-marker # remains a # in the
URL, etc. So, it would seem as though simple URL encoding should be
enough to handle pound signs, even given that we use escaping for them
in the `.expl` format.

However, as with underscores, we have a behavior present in external
services that prevents this scheme from being sufficient. Namely. some
services like Gmail encode pound signs after the first, and thus
http://example.com/A#B/C#D would automatically become
http://example.com/A#B/C%23D. This creates an ambiguity at the decoding
stage, ie is `C%23D` one topic, `C#D`, or is it the topic `C` and the
subtopic `D`? The automatic encoding creates a collision which makes
decoding the URL impossible.

So, in such a case, our general approach has been to use encoded
backslashes to encode characters in situations where simple encoding is
ambiguous due to potential collisions, as with underscore. So, our
resulting pound encoding schema would be:

`#`     -  this in a URL is a fragment id/subtopic marker ie A#B/C#D
`%23`   - also a # that is a fragment identifier in the path structure
`%5C#`  - no string should produce this URL but is parsed as topic
          name # `%5C%23` - also a # that is part of a topic or subtopic
          name
`%2523` - a literal "%23" that is three letters of a topic name

Because not all characters are encoded, not all characters should be
decoded, so this commit introduces a new rule-based encoding and
decoding scheme. To ensure that the output of earlier rules are not
input to later rules, we use a recursive string subdivision approach
wherein upon finding a match, the string is divided, subportions
processed for further rules, and then the portions are joined with the
given replacement so that the replacement is not itself replaced by any
lateral rule, eg ' ' becoming '_' and then becoming '%5C_', etc.

Chrome encodes certain special characters in URLs automatically, namely
backtick \`, carrot ^, and double quote ", (although not single quote.)
Ie, these characters if typed into the URL bar will produce a
window.location that has encoded characters. As such, we have to
include these characters on the `urlEncodingRules` list so that we know
to decode them when we see them. However, we don't need to employ
escaping because that was only necessary regarding '_' because the
encoded form is automatically _decoded_ by Chrome, and regarding '#'
because of the special topic-name-# vs. path-structure-# ambiguity in
Canopy's business logic. Besides those two examples, any character can
be identified by simple encoding even where the browser automatically
encodes it, so long as the browser is not automatically _decoding_ as
it does for underscore.