Add padding around blockquote break tags
Now, we replace break tags in block quotes and give them padding. It
wasn't possible to just break block quote contents on newlines from the
original text and put them in separate top-level spans because it is
permitted to have bold or italic elements that enclose multiple 
paragraphs in the block quote, so we need a solution that is compatible
with any intra-blockquote DOM hierarchy.