Refactor upward movement to link off screen
Upward movement is affected by code in the arrow key handler and the
shouldAnimate function - the arrow key handler decides whether to
updateView or whether to just scroll in anticipation of a second key-
press, whereas shouldAnimate decides whether a given updateView should
animate. Here, we adopt the view that if a new target is high enough
above the viewport to merit animation on that basis, it merits an
anticipatory scroll instead of updating the selection at all. So,
the remaining considerations in deciding whether to animate are merely
if the new target is much higher, whether on or off screen, and whether
the movement is two-part. If two-part, the animation code itself will
decide whether the first destination is already focused, so as to cancel
the first movement.