Bug fix: Live insertion of intermediary paragraphs breaks render
A recent performance refactor parallelized the render of path segments
by breaking the path into pieces, requesting JSON and rendering topic
subtrees separately, and only then combining them into one tree.

One downside of this approach as compared to initiating the render of
a path component upon encountering the global reference that predicates
the path subcomponent is that it is impossible to render subtrees in
error due to a URL path that has adjacency not corresponding to parent
links in the actual data.

The new approach creates the following possibility: a user may be
viewing a page like https://example.com/A#B/C, and then the data on the
server may change so that B is no longer the subtopic of A that contains
the global reference to C, but rather B's child B1. Previously in such
a case, we were rendering C and registering it in the Paragraph class's
instance store as `A#B/C`, even though there was no such connected path,
and then in displayPath we were merely checking if there was a paragraph
in the store by that name, never checking if in fact that paragraph had
been successfully subsumed from the root ie representing a valid path
for display.

We can solve this issue by making a very small change to the code.
Previously, subtopicElementPromise ie the subtopic of path segment A
became the parentElementPromise for segment B, and thus B's appending
occurred narrowly upon A's and B's render, regardless of A's ancestry.

Now, subtopicElementPromise ie parentElementPromise in n+1 does not
resolve until subtopicElement's own parentElementPromise has resolved.
Thus, appending of B to A waits for A's parentElementPromise, which now
waits for its parent, etc until you hit an immediately resolving parent
ie canopyContainer or an existing DOM node. Thus the only elements that
are successfully appended and registered are those that have been
successfully subsumed.