Clarify hyperlink regex behavior
There are a number of challenges inherent in designing a rule to match
hyperlinks and image links such that nested images are matched without
special indication. Ie, [![a](b)](c) should become a linked image merely
because "![a](b)" is the link's text, and "c" is its URL, and most text
fields in Canopy are reparsed for subtokens like bold, italic, or links.

These are some of the cases that present challenges:

- [link 1](http://google.com) [link 2](http://google.com)

Here we need to recognize that these are two links, not one large link
with the text "link 1](http://google.com) [link 2" which is the default
behavior if our regex is greedy. This case would be solved if the link
text was not greedy.

- [![pic](https://en.wikipedia.org/favicon.ico)](en.wikipedia.org/)

This is a nested image. We have to clarify why this should be a
hyperlink with text "![pic](https://en.wikipedia.org/favicon.ico" and
not a hyperlink with text "![pic" and URL
"https://en.wikipedia.org/favicon.ico)](en.wikipedia.org/" - here, we
need the link text regex to be greedy.

Additionally, any escaping that might indicate that the image's brackets
and parenthases are not part of the link will cause problems when we
are parsing the subfields and need those brackets and parens to be
unescaped - removing single backslashes in general between link and
image parsing would remove other intentional escaping unintentionally,
eg special characters in the image description, requiring double
escaping for those characters specifically, etc.

- [![pic](a.jpeg)![pic](a.jpeg)](https://en.wikipedia.org/)

This is a double nested image, so however we recognize nested images,
it can't be in a one-off fixed rule.

- This is a nested image red herring [![abc](https://en.wikipedia.org/)

However we recognize nested images it can't be too fixedly.

- ([link](a.com)) and an escaped one ([link](a.com)\)

These two links present an ambiguity of whether the final paren is
plaintext after the link or whether it is part of the link's URL.

- [[link](http://google.com)](http://google.com)

This is a double nested link, and it probably shouldn't get parsed as
such.

====

So, given all these constraints, the following approach is used.

Hyperlink text is non-greedy, thus two hyperlinks side-by-side are not
merged. However, non-greedy parsing makes nested images difficult due
to the abundance of brackets and parens. So, now if we see the text
sequence `![` in the text of a hyperlink, this causes us to go into
a greedy mode until we either see an unescaped `)`, or until it becomes
clear we are not in the text field of a hyperlink. That way, we try
to make the entire nested image or images into the text of the current
hyperlink, but if that fails as in the red herring case, we fall back to
a regular non-greedy interpretation. URL match is also non-greedy 
because it is easier to escape a URL paren than it is to get the link to
stop early, which could be done by escaping characters after ) which
look URL-like.