Encoding non-ascii characters fix
Previously, we established an encoding scheme that handled conversions
of topic names into URLs, with custom handling of spaces being converted
to underscores, and differentiating between # signs that were part of
topic names versus those which were path separators, with an
understanding that some browsers/services automatically encoded # to
%23, making a %23 potentially ambiguous as to whether it was originally
a name character or an accidentally encoded path separator, etc.

This previous encoding scheme handled all these cases, however, it did
not account for the fact that many ascii characters and all non-ascii
characters get automatically URL encoded by browsers, and in a sometimes
platform-dependent manner. Because URL conversion must be decodable,
it is important to know exactly which characters were encoded and which
are literal %XY sequences. Therefore, we had to augment our encoding
scheme to be able to keep track of which characters have been encoded
by the browser. Now, the obvious approach would be to URL encode all
characters, so that the browser cannot surprise us with a conversion.
However, it is a design goal that URLs be as human-readable as possible,
and even though some encoded characters are displayed decoded in the URL
bar, if we encode a character like @ that some browsers do not encode,
on those browsers, the encoding will seem intentional and will be
displayed as such.

So, we need a new encoding scheme that handles the special behavior of
_ and # in Canopy while at the same time separating clearly between
characters that were browser-encoded and those that are %XY literals,
while at the same time not encoding characters that don't need to be
encoded, on a per-platform basis. (And not only per-flatform, bu also
per path modality, as even for a given browser, characters which are
allowed in the path may not be the same as those allowed in the hash,
and so a Canopy site using hash URLs may have different automatic
encodings than the same site and browser using path URLs and a Canopy
node server.)

So, the new approach is to do a URI decode after all Canopy conversion
business logic has completed, followed by the removal of single
backslashes once they have been used during prior stages to identify
literals. This way, we can put characters like @ and } in the URL,
and each browser can decide which to encode, and so long as we always
encode %, the entire single-encoding-space is reserved for automatic
conversions, and the decode we do will undo all of them while returning
%XY sequences that were literal to their original form.

Below is a table that maps out the types of characters that are found
in topic names, and the sequence of operations that are necessary to
convert them to their next form. Doing the operations in the wrong
order may cause collisions, such as decoding backslashes that have not
yet been used to differentiate between path-# and topic-name-#
characters, etc. Because special characters like _ literal and
topic-name-# are encoded as %5C# or %5C_, we do not need to reverse the
entire encoding process, as decoding to \# and \_ followed by deescaping
ie removing single backslashes suffices.

+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|   (Display)   |  %  | Subtopic | Subtopic |   \#   |  \#  |  "  |    א   | ⎵ | Style _ |  \_  |  %23  | Topic |  /  |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|   Mixed Case  |  %  | Subtopic | Subtopic |    #   |   #  |  "  |    א   | ⎵ | Nothing |   _  |  %23  | Topic |  /  |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
| Topic Convert | %25 | Subtopic | Subtopic | %5C%23 | %5C# |  "  |    א   | _ | Nothing | %5C_ | %2523 | Topic | %2F |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|  Path Convert | %25 |     #    |    %23   | %5C%23 | %5C# |  "  |    א   | _ | Nothing | %5C_ | %2523 |   /   | %2F |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|    Browser    | %25 |     #    |    %23   | %5C%23 | %5C# | %22 | %D7%90 | _ | Nothing | %5C_ | %2523 |   /   | %2F |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|     (URL)     |     |          |          |        |      |     |        |   |         |      |       |       |     |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|               |     |          |          |        |      |     |        |   |         |      |       |       |     |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|      URL      | %25 |     #    |    %23   | %5C%23 | %5C# | %22 | %D7%90 | _ | Nothing | %5C_ | %2523 |   /   | %2F |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|  Path Convert | %25 | Subtopic | Subtopic | %5C%23 | %5C# | %22 | %D7%90 | _ | Nothing | %5C_ | %2523 | Topic | %2F |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
| Topic Convert | %25 | Subtopic | Subtopic | %5C%23 | %5C# | %22 | %D7%90 | ⎵ | Nothing | %5C_ | %2523 | Topic | %2F |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|     Decode    |  %  | Subtopic | Subtopic |   \#   |  \#  |  "  |    א   | ⎵ | Nothing |  \_  |  %23  | Topic |  /  |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|    Unescape   |  %  | Subtopic | Subtopic |    #   |   #  |  "  |    א   | ⎵ | Nothing |   _  |  %23  | Topic |  /  |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+
|  (Mixed Case) |     |          |          |        |      |     |        |   |         |      |       |       |     |
+---------------+-----+----------+----------+--------+------+-----+--------+---+---------+------+-------+-------+-----+