Package live rebuild code for reuse
Previously the playground code contained both the logic for setting up
the playground containers and also the logic for running the Canopy
parser live in the browser. Now, there is a new webpack endpoint called
rebuild_canopy that puts a function on the window object called
`rebuildCanopy` which takes a bulk file string argument. Users can
create any editor or system of managing bulk files, and then call this
function with the result to rebuild the `#_canopy` element in the DOM.

The rebuildCanopy function will rethrow errors from the parser, so
consumers can wrap it in a try-catch block and decide how they want
to present them.