HTML token insertions
Previously, if a user inserted raw HTML into a paragraph, there was no
way of allowing Canopy links and other types of tokens to be inserted.
This might be desirable, for example, if the user wants to make an
HTML table with custom styling, but to insert regular Canopy links
into the table cells.

This is an example of the new format:
```
<div style="background-color: green">
  This is a link: {{[[New York]]}}
</div>
<div>
  These are not: \{{[[New York]]}} {\{[[New York]]}} {{[[New York]]\}} {{[[New York]]}\}
</div>
```