Fix non-extant link error in Canopy Notebook
Because Canopy Notebook creates an h1 outside of the app, our test
for page load was falsely positive, and so we were never finding
links from metadata on initial page load.

Now, we put a class on Canopy h1s to distinguish them from other
headers on the page, but also we no longer throw an error for
invalid link metadata because this is common to occur during the
writing process, and so we should just silently default eg to the
path's parent link.