Replace textline token with text token newline rendering
Initially newlines in expl files were ignored, but the desire for multi-
line paragraphs led to rendering expl newlines and introducing padding.
This led to the need to identify where lines of text ended to insert
newlines and padding. Initially a text-line matcher was introduced that
would greedily match text until a newline. The problem was that this
could fail to give other token types the chance to be triggered such
the newline would end up being text within another token type. (Although
no multiline token type at present may begin within the middle of a
line, one possible application would be writing a tool tip and using
real newlines as literals.)

So, the new approach is to return to the original method of matching
text with buffer after exhausting other matchers, and introducing the
desired behavior at render time, ie, lines of text with newlines and
padding being not a new type of token, but a kind of conditional
rendering of text tokens generally.

In this scheme, text tokens with newlines trigger the creation of a
div.canopy-text-container node, and newlines inside this container type
if a direct child of p.canopy-paragraph get padding, to avoid giving
padding necessarily to multiline text within other token types, like
tool tips.